requirements.txt
Flask<=2.3.2
Werkzeug<=2.3.6
openpyxl==3.1.2
Pillow==11.2.1
watchdog==2.2.0
gunicorn==23.0.0
